refactor(suite-native): remove redundant bottom sheet
