+++ suite-native/intl/src/messages.ts
@@ -1796,16 +1796,6 @@ export const messages = {
             },
         },
     },
-    moduleLabeling: {
-        exportBip329BottomSheet: {
-            title: 'Export labels',
-            button: 'Export labels',
-            description: 'Export address and output labels form this account in BIP 329 format.',
-        },
-        exportSuccessfulToast: 'Labels exported.',
-        exportFailedToast: 'Failed to export labels.',
-        fileSavingNotSupported: 'File saving is not supported on this platform.',
-    },
     moduleAccounts: {
         accountNotFound: 'Account {accountKey} not found.',
         tokens: {
@@ -1854,7 +1844,10 @@ export const messages = {
             title: 'Export labels',
             button: 'Export',
             description:
-                'Export your transaction labels for this account to a .jsonl file. You can import this file into other wallet apps that support label import.',
+                'Export transaction labels for this account. Compatible with other wallet applications that support the BIP-329 format.',
+            exportSuccessfulToast: 'Labels exported.',
+            exportFailedToast: 'Failed to export labels.',
+            fileSavingNotSupported: 'File saving is not supported on this platform.',
         },
     },
     transactions: {
