+++ suite-native/labeling/tsconfig.json
@@ -2,7 +2,6 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
-        { "path": "../../suite-common/bip329" },
         { "path": "../../suite-common/device" },
         {
             "path": "../../suite-common/message-system"
@@ -41,9 +40,6 @@
         { "path": "../settings" },
         { "path": "../toasts" },
         { "path": "../../packages/connect" },
-        {
-            "path": "../../packages/styles-native"
-        },
         { "path": "../../packages/type-utils" },
         { "path": "../test-utils-store" }
     ]
