+++ suite-native/module-accounts-management/package.json
@@ -53,12 +53,16 @@
         "@suite-native/navigation": "workspace:*",
         "@suite-native/qr-code": "workspace:*",
         "@suite-native/services": "workspace:*",
+        "@suite-native/toasts": "workspace:*",
         "@suite-native/tokens": "workspace:*",
         "@suite-native/transactions": "workspace:*",
         "@trezor/connect": "workspace:*",
         "@trezor/styles-native": "workspace:*",
+        "@trezor/type-utils": "workspace:*",
         "@trezor/utils": "workspace:*",
         "date-fns": "^4.1.0",
+        "expo-file-system": "~55.0.10",
+        "expo-sharing": "~55.0.11",
         "jotai": "2.17.1",
         "react": "19.2.4",
         "react-hook-form": "^7.71.2",
