+++ suite-native/module-accounts-management/tsconfig.json
@@ -54,12 +54,14 @@
         { "path": "../navigation" },
         { "path": "../qr-code" },
         { "path": "../services" },
+        { "path": "../toasts" },
         { "path": "../tokens" },
         { "path": "../transactions" },
         { "path": "../../packages/connect" },
         {
             "path": "../../packages/styles-native"
         },
+        { "path": "../../packages/type-utils" },
         { "path": "../../packages/utils" },
         { "path": "../test-utils-store" },
         { "path": "../../packages/requirements" }
